# contracts/openapi.json
{
  "components": {
    "schemas": {
      "Body_submit_report_api_v1_reports_post": {
        "properties": {
          "consent": {
            "const": true,
            "title": "Consent",
            "type": "boolean"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 20,
            "title": "Description",
            "type": "string"
          },
          "language_hint": {
            "enum": [
              "en",
              "kn",
              "hi",
              "mixed"
            ],
            "title": "Language Hint",
            "type": "string"
          },
          "locality_label": {
            "maxLength": 160,
            "minLength": 3,
            "title": "Locality Label",
            "type": "string"
          },
          "synthetic_demo_confirmation": {
            "const": true,
            "title": "Synthetic Demo Confirmation",
            "type": "boolean"
          },
          "voice": {
            "anyOf": [
              {
                "contentMediaType": "application/octet-stream",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Voice"
          }
        },
        "required": [
          "description",
          "language_hint",
          "locality_label",
          "consent",
          "synthetic_demo_confirmation"
        ],
        "title": "Body_submit_report_api_v1_reports_post",
        "type": "object"
      },
      "CandidatePrerequisite": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "state": {
            "enum": [
              "unknown",
              "required",
              "satisfied",
              "not_applicable"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "code",
          "state"
        ],
        "title": "CandidatePrerequisite",
        "type": "object"
      },
...
      "EvidenceItem": {
        "additionalProperties": false,
        "properties": {
          "classification": {
            "enum": [
              "real_public",
              "derived_public",
              "synthetic_demo"
            ],
            "title": "Classification",
            "type": "string"
          },
          "contributes_to_rating": {
            "title": "Contributes To Rating",
            "type": "boolean"
          },
          "geography": {
            "title": "Geography",
            "type": "string"
          },
          "geography_kind": {
            "title": "Geography Kind",
            "type": "string"
          },
          "geography_vintage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geography Vintage"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "prohibited_inferences": {
            "items": {
              "type": "string"
            },
            "title": "Prohibited Inferences",
            "type": "array"
          },
          "reference_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Period"
          },
          "role": {
            "enum": [
              "local_decision_evidence",
              "broader_context"
            ],
            "title": "Role",
            "type": "string"
          },
          "semantics": {
            "title": "Semantics",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "supported_inferences": {
            "items": {
              "type": "string"
            },
            "title": "Supported Inferences",
            "type": "array"
          },
          "transformation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transformation"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "key",
          "label",
          "classification",
          "kind",
          "value",
          "unit",
          "geography",
          "geography_kind",
          "geography_vintage",
          "reference_period",
          "semantics",
          "source_url",
          "transformation",
          "contributes_to_rating",
          "role",
          "supported_inferences",
          "prohibited_inferences"
        ],
        "title": "EvidenceItem",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IncidentDetail": {
        "additionalProperties": false,
        "properties": {
          "data_version": {
            "title": "Data Version",
            "type": "string"
          },
          "decision_boundary": {
            "title": "Decision Boundary",
            "type": "string"
          },
          "incident": {
            "$ref": "#/components/schemas/IncidentSummary"
          },
          "relationship_explanation": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Relationship Explanation",
            "type": "object"
          },
          "reports": {
            "items": {
              "$ref": "#/components/schemas/ReportEvidence"
            },
            "title": "Reports",
            "type": "array"
          }
        },
        "required": [
          "data_version",
          "incident",
          "reports",
          "relationship_explanation",
          "decision_boundary"
        ],
        "title": "IncidentDetail",
        "type": "object"
      },
      "IncidentSummary": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "classification": {
            "const": "synthetic_demo",
            "title": "Classification",
            "type": "string"
          },
          "event_start": {
            "format": "date-time",
            "title": "Event Start",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "relationship_state": {
            "title": "Relationship State",
            "type": "string"
          },
          "report_count": {
            "title": "Report Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "title",
          "category",
          "locality",
          "event_start",
          "report_count",
          "relationship_state",
          "classification"
        ],
        "title": "IncidentSummary",
        "type": "object"
      },
...
      "NeedSummary": {
        "additionalProperties": false,
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "classification": {
            "const": "synthetic_demo",
            "title": "Classification",
            "type": "string"
          },
          "geography": {
            "title": "Geography",
            "type": "string"
          },
          "geography_kind": {
            "title": "Geography Kind",
            "type": "string"
          },
          "geography_membership_method": {
            "title": "Geography Membership Method",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "incident_count": {
            "title": "Incident Count",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "lifecycle": {
            "title": "Lifecycle",
            "type": "string"
          },
          "priority_band": {
            "title": "Priority Band",
            "type": "string"
          },
          "report_count": {
            "title": "Report Count",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "title",
          "category",
          "geography",
          "geography_kind",
          "geography_membership_method",
          "lifecycle",
          "incident_count",
          "report_count",
          "priority_band",
          "classification"
        ],
        "title": "NeedSummary",
        "type": "object"
      },
      "NeedWorkspace": {
        "additionalProperties": false,
        "properties": {
          "alternative_hypotheses": {
            "items": {
              "type": "string"
            },
            "title": "Alternative Hypotheses",
            "type": "array"
          },
          "candidate": {
            "$ref": "#/components/schemas/ProjectCandidate"
          },
          "data_version": {
            "title": "Data Version",
            "type": "string"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "evidence_notice": {
            "title": "Evidence Notice",
            "type": "string"
          },
          "human_decision_notice": {
            "title": "Human Decision Notice",
            "type": "string"
          },
          "hypothesis": {
            "title": "Hypothesis",
            "type": "string"
          },
          "hypothesis_label": {
            "const": "Suspected Civic Need",
            "title": "Hypothesis Label",
            "type": "string"
          },
          "incidents": {
            "items": {
              "$ref": "#/components/schemas/IncidentSummary"
            },
            "title": "Incidents",
            "type": "array"
          },
          "need": {
            "$ref": "#/components/schemas/NeedSummary"
          },
          "priority": {
            "$ref": "#/components/schemas/PrioritySummary"
          }
        },
        "required": [
          "data_version",
          "need",
          "hypothesis",
          "hypothesis_label",
          "alternative_hypotheses",
          "incidents",
          "evidence",
          "priority",
          "candidate",
          "evidence_notice",
          "human_decision_notice"
        ],
        "title": "NeedWorkspace",
        "type": "object"
      },
...
      "PriorityComponent": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "rating": {
            "anyOf": [
              {
                "maximum": 3.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          }
        },
        "required": [
          "code",
          "label",
          "rationale",
          "evidence_refs"
        ],
        "title": "PriorityComponent",
        "type": "object"
      },
      "PrioritySummary": {
        "additionalProperties": false,
        "properties": {
          "abstention_codes": {
            "items": {
              "type": "string"
            },
            "title": "Abstention Codes",
            "type": "array"
          },
          "band": {
            "enum": [
              "high",
              "moderate",
              "lower",
              "not_comparable"
            ],
            "title": "Band",
            "type": "string"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/PriorityComponent"
            },
            "title": "Components",
            "type": "array"
          },
          "eligible": {
            "title": "Eligible",
            "type": "boolean"
          },
          "policy_notice": {
            "title": "Policy Notice",
            "type": "string"
          },
          "sensitivity_profiles": {
            "items": {
              "type": "string"
            },
            "title": "Sensitivity Profiles",
            "type": "array"
          },
          "sensitivity_status": {
            "enum": [
              "stable",
              "sensitive",
              "not_comparable"
            ],
            "title": "Sensitivity Status",
            "type": "string"
          }
        },
        "required": [
          "eligible",
          "band",
          "components",
          "sensitivity_status",
          "sensitivity_profiles",
          "abstention_codes",
          "policy_notice"
        ],
        "title": "PrioritySummary",
        "type": "object"
      },
      "ProjectCandidate": {
        "additionalProperties": false,
        "properties": {
          "catalogue_key": {
            "title": "Catalogue Key",
            "type": "string"
          },
          "catalogue_version": {
            "title": "Catalogue Version",
            "type": "string"
          },
          "conditional_wording": {
            "title": "Conditional Wording",
            "type": "string"
          },
          "decision_boundary": {
            "title": "Decision Boundary",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "lifecycle": {
            "title": "Lifecycle",
            "type": "string"
          },
          "prerequisites": {
            "items": {
              "$ref": "#/components/schemas/CandidatePrerequisite"
            },
            "title": "Prerequisites",
            "type": "array"
          },
          "uncertainty": {
            "items": {
              "type": "string"
            },
            "title": "Uncertainty",
            "type": "array"
          },
          "works_overlap_evidence": {
            "items": {
              "type": "string"
            },
            "title": "Works Overlap Evidence",
            "type": "array"
          },
          "works_overlap_outcome": {
            "title": "Works Overlap Outcome",
            "type": "string"
          }
        },
        "required": [
          "id",
          "catalogue_key",
          "catalogue_version",
          "label",
          "conditional_wording",
          "lifecycle",
          "works_overlap_outcome",
          "works_overlap_evidence",
          "prerequisites",
          "uncertainty",
          "decision_boundary"
        ],
        "title": "ProjectCandidate",
        "type": "object"
      },
...
      "ReceiptStatus": {
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "analysis_class": {
            "enum": [
              "pending",
              "stored_sample",
              "fresh_fixture",
              "fresh_ai"
            ],
            "title": "Analysis Class",
            "type": "string"
          },
          "analysis_state": {
            "title": "Analysis State",
            "type": "string"
          },
          "evidence_class": {
            "const": "synthetic_demo",
            "title": "Evidence Class",
            "type": "string"
          },
          "generalized_area": {
            "title": "Generalized Area",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "received",
              "analysing",
              "processed",
              "needs_review"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "public_id",
          "status",
          "accepted_at",
          "generalized_area",
          "evidence_class",
          "analysis_class",
          "analysis_state",
          "message"
        ],
        "title": "ReceiptStatus",
        "type": "object"
      },
      "ReportAccepted": {
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          },
          "status": {
            "const": "received",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "public_id",
          "status",
          "accepted_at",
          "message"
        ],
        "title": "ReportAccepted",
        "type": "object"
      },
      "ReportEvidence": {
        "additionalProperties": false,
        "properties": {
          "accepted_at": {
            "format": "date-time",
            "title": "Accepted At",
            "type": "string"
          },
          "classification": {
            "const": "synthetic_demo",
            "title": "Classification",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "interpretation_classification": {
            "const": "ai_derived",
            "title": "Interpretation Classification",
            "type": "string"
          },
          "interpretation_summary": {
            "title": "Interpretation Summary",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "locality": {
            "title": "Locality",
            "type": "string"
          },
          "original_text": {
            "title": "Original Text",
            "type": "string"
          },
          "public_id": {
            "title": "Public Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "public_id",
          "original_text",
          "language",
          "accepted_at",
          "locality",
          "interpretation_summary",
          "classification",
          "interpretation_classification"
        ],
        "title": "ReportEvidence",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Explainable civic needs-to-project decision intelligence.",
    "title": "CivicLens API",
    "version": "dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/officer/incidents/{incident_id}": {
      "get": {
        "operationId": "incident_detail_api_v1_officer_incidents__incident_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Incident Detail",
        "tags": [
          "officer-demo"
        ]
      }
    },
...
    "/api/v1/officer/needs/{need_id}": {
      "get": {
        "operationId": "need_workspace_api_v1_officer_needs__need_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "need_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Need Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NeedWorkspace"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Need Workspace",
        "tags": [
          "officer-demo"
        ]
      }
    },
...
    "/api/v1/receipts/{public_id}": {
      "get": {
        "operationId": "receipt_status_api_v1_receipts__public_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "public_id",
            "required": true,
            "schema": {
              "title": "Public Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Receipt-Capability",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 32,
              "title": "X-Receipt-Capability",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Receipt Status",
        "tags": [
          "report-intake"
        ]
      }
    },
    "/api/v1/reports": {
      "post": {
        "operationId": "submit_report_api_v1_reports_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 16,
              "title": "Idempotency-Key",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Receipt-Capability",
            "required": true,
            "schema": {
              "maxLength": 256,
              "minLength": 32,
              "title": "X-Receipt-Capability",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_report_api_v1_reports_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Submit Report",
        "tags": [
          "report-intake"
        ]
      }
    }
  }
}